requirements.txt
telethon
openai
fpdf
python-dotenv